+++ src/mako/chrono/TimeInterface.php
@@ -162,14 +162,4 @@ interface TimeInterface extends DateTimeInterface
 	 * @return int
 	 */
 	public function daysInMonth(): int;
-
-	/**
-	 * Returns a formatted date string according to current locale settings.
-	 *
-	 * @deprecated
-	 *
-	 * @param  string $format Date format
-	 * @return string
-	 */
-	public function formatLocalized(string $format): string;
 }