+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The major version bump is due to dropped support for PHP `8.0` and a several bre
 * Removed the NuoDB query compiler.
 * Removed the deprecated event library (use the bus library instead).
 * Removed the deprecated command bus library (use the bus library instead).
+* Removed the deprecated TimeInterface::formatLocalized() method.
 
 --------------------------------------------------------